Downloads multiple Parameter Configurations from the mlrHyperopt servers

Retrieve the Paramater Configurations for the given ids from the mlrHyperopt servers.

downloadParConfigs(ids)

Arguments

ids

[character] One ore more unique identifiers of the Parameter Set

Value

[List of ParConfigs]

Examples

par.configs = downloadParConfigs(c("8","29"))
print(par.configs)

#> [[1]]
#> $par.set
#>            Type len Def            Constr Req Tunable Trafo
#> cost    numeric   -   -           0 to 15   -    TRUE     -
#> kernel discrete   -   - polynomial,radial   -    TRUE     -
#> degree  integer   -   3          1 to Inf   Y    TRUE     -
#> gamma   numeric   -   -           -5 to 5   -    TRUE     Y
#> 
#> $par.vals
#> $par.vals$cachesize
#> [1] 100
#> 
#> $par.vals$tolerance
#> [1] 0.01
#> 
#> 
#> $associated.learner.class
#> [1] "classif.svm"
#> 
#> $associated.learner.type
#> [1] "classif"
#> 
#> $associated.learner.name
#> [1] "svm"
#> 
#> attr(,"class")
#> [1] "ParConfig"
#> 
#> [[2]]
#> $par.set
#>          Type len Def    Constr Req Tunable Trafo
#> cost  numeric   -   - -15 to 15   -    TRUE     Y
#> gamma numeric   -   - -15 to 15   -    TRUE     Y
#> 
#> $par.vals
#> list()
#> 
#> $associated.learner.class
#> [1] "classif.svm"
#> 
#> $associated.learner.type
#> [1] "classif"
#> 
#> $associated.learner.name
#> [1] "svm"
#> 
#> attr(,"class")
#> [1] "ParConfig"
#>