Generates a hyperparameter tuning control object

Tries to automatically create a suitable hyperparameter tuning control.

generateHyperControl(task, par.config = NULL, learner = NULL,
  budget.evals = 25)

Arguments

task

[Task] The Task

par.config

[ParConfig] The Parameter Configuration

learner

[Learner] Learner

budget.evals

[integer] How many function evaluations do you want to allow? Default is 25.

Value

[HyperControl]

Examples

par.config = getDefaultParConfig("regr.randomForest")
hyper.control = generateHyperControl(bh.task, par.config)
# get what is inside
getHyperControlMeasures(hyper.control)

#> [[1]]
#> Name: Mean of squared errors
#> Performance measure: mse
#> Properties: regr,req.pred,req.truth
#> Minimize: TRUE
#> Best: 0; Worst: Inf
#> Aggregated by: test.mean
#> Note: Defined as: mean((response - truth)^2)
#> 
#> [[2]]
#> Name: Time of fitting the model
#> Performance measure: timetrain
#> Properties: classif,classif.multi,multilabel,regr,surv,costsens,cluster,req.model
#> Minimize: TRUE
#> Best: 0; Worst: Inf
#> Aggregated by: test.mean
#> Note: 
#> 
#> [[3]]
#> Name: Time of predicting test set
#> Performance measure: timepredict
#> Properties: classif,classif.multi,multilabel,regr,surv,costsens,cluster,req.pred
#> Minimize: TRUE
#> Best: 0; Worst: Inf
#> Aggregated by: test.mean
#> Note: 
#> 

getHyperControlMlrControl(hyper.control)

#> Tune control: TuneControlMBO
#> Same resampling instance: TRUE
#> Imputation value: <worst>
#> Start: <NULL>
#> 
#> Tune threshold: FALSE
#> Further arguments: 

getHyperControlResampling(hyper.control)

#> Resample description: OOB bootstrapping with 5 iterations.
#> Predict: test
#> Stratification: FALSE

# change what is inside
hyper.control = setHyperControlMeasures(hyper.control, measures = medse)
hyper.control = setHyperControlMlrControl(
  hyper.control,
  mlr.control = makeTuneControlRandom(maxit = 10))
hyper.control = setHyperControlResampling(hyper.control, resampling = cv3)
hyperopt(task = bh.task, par.config = par.config, hyper.control = hyper.control)

#> [Tune] Started tuning learner regr.randomForest for parameter set:

#>             Type len  Def    Constr Req Tunable Trafo
#> ntree    numeric   - 5.64 0 to 6.64   -    TRUE     Y
#> nodesize integer   -    1   1 to 10   -    TRUE     -
#> mtry     integer   -    4   1 to 13   -    TRUE     -

#> With control class: TuneControlRandom

#> Imputation value: Inf

#> [Tune-x] 1: ntree=331; nodesize=3; mtry=11

#> [Tune-y] 1: medse.test.mean=2.69; time: 0.0 min

#> [Tune-x] 2: ntree=65; nodesize=9; mtry=12

#> [Tune-y] 2: medse.test.mean= 2.8; time: 0.0 min

#> [Tune-x] 3: ntree=16; nodesize=6; mtry=13

#> [Tune-y] 3: medse.test.mean=   3; time: 0.0 min

#> [Tune-x] 4: ntree=502; nodesize=5; mtry=5

#> [Tune-y] 4: medse.test.mean=2.72; time: 0.0 min

#> [Tune-x] 5: ntree=47; nodesize=3; mtry=6

#> [Tune-y] 5: medse.test.mean= 2.5; time: 0.0 min

#> [Tune-x] 6: ntree=106; nodesize=2; mtry=5

#> [Tune-y] 6: medse.test.mean=2.58; time: 0.0 min

#> [Tune-x] 7: ntree=431; nodesize=8; mtry=7

#> [Tune-y] 7: medse.test.mean=2.78; time: 0.0 min

#> [Tune-x] 8: ntree=838; nodesize=2; mtry=10

#> [Tune-y] 8: medse.test.mean= 2.5; time: 0.1 min

#> [Tune-x] 9: ntree=495; nodesize=7; mtry=4

#> [Tune-y] 9: medse.test.mean=2.61; time: 0.0 min

#> [Tune-x] 10: ntree=680; nodesize=5; mtry=3

#> [Tune-y] 10: medse.test.mean=2.44; time: 0.0 min

#> [Tune] Result: ntree=680; nodesize=5; mtry=3 : medse.test.mean=2.44

#> Tune result:
#> Op. pars: ntree=680; nodesize=5; mtry=3
#> medse.test.mean=2.44