Create a dictionary based on the task.

Returns a dictionary, which contains the Task itself (task), the number of features (p) the model is trained on, the number of observations (n.task) of the task in general, the number of observations (n) in the current subset, the task type (type) and in case of classification tasks, the number of class levels (k) in the general task.

getTaskDictionary(task)

Arguments

task
[Task] The Task

Value

[list]. Used for evaluating the expressions within a parameter, parameter set or list of parameters.

Examples

task = makeClassifTask(data = iris, target = "Species")
getTaskDictionary(task)

#> $task
#> Supervised task: iris
#> Type: classif
#> Target: Species
#> Observations: 150
#> Features:
#>    numerics     factors     ordered functionals 
#>           4           0           0           0 
#> Missings: FALSE
#> Has weights: FALSE
#> Has blocking: FALSE
#> Has coordinates: FALSE
#> Classes: 3
#>     setosa versicolor  virginica 
#>         50         50         50 
#> Positive class: NA
#> 
#> $p
#> [1] 4
#> 
#> $n.task
#> [1] 150
#> 
#> $type
#> [1] "classif"
#> 
#> $n
#> [1] 150
#> 
#> $k
#> [1] 3
#> 
#> $t
#>     setosa versicolor  virginica 
#>         50         50         50 
#>