Hyperparameter Tuning Control Object

Defines how the hyperparameter tuning should be conducted

makeHyperControl(
  mlr.control = NULL,
  resampling = NULL,
  measures = NULL,
  par.config = NULL
)

Arguments

mlr.control
[TuneControl]
Control object for search method. Also selects the optimization algorithm for tuning.

resampling
[ResampleDesc] The resampling determines how the performance is obtained during tuning.

measures
[Measure | list of Measure]
Performance measure(s) to evaluate. Default is the default measure for the task, see here getDefaultMeasure.

par.config
[ParConfig] The Parameter Configuration

Value

[HyperControl]

See also

Examples

hyper.control = makeHyperControl(
  mlr.control = makeTuneControlRandom(maxit = 5),
  resampling = cv2,
  measures = acc
)
hyperopt(task = iris.task, learner = "classif.svm", hyper.control = hyper.control)

#> [Tune] Started tuning learner classif.svm for parameter set:

#>          Type len Def    Constr Req Tunable Trafo
#> cost  numeric   -   0 -15 to 15   -    TRUE     Y
#> gamma numeric   -  -2 -15 to 15   -    TRUE     Y

#> With control class: TuneControlRandom

#> Imputation value: -0

#> [Tune-x] 1: cost=0.00103; gamma=4.7e+03

#> [Tune-y] 1: acc.test.mean=0.2333333; time: 0.0 min

#> [Tune-x] 2: cost=4.28e-05; gamma=3.21e+04

#> [Tune-y] 2: acc.test.mean=0.2333333; time: 0.0 min

#> [Tune-x] 3: cost=0.016; gamma=0.000804

#> [Tune-y] 3: acc.test.mean=0.2333333; time: 0.0 min

#> [Tune-x] 4: cost=0.158; gamma=3.27e+03

#> [Tune-y] 4: acc.test.mean=0.2333333; time: 0.0 min

#> [Tune-x] 5: cost=0.000112; gamma=6.94

#> [Tune-y] 5: acc.test.mean=0.2333333; time: 0.0 min

#> [Tune] Result: cost=0.158; gamma=3.27e+03 : acc.test.mean=0.2333333

#> Tune result:
#> Op. pars: cost=0.158; gamma=3.27e+03
#> acc.test.mean=0.2333333