Uploads a Parameter Set to the mlrHyperopt servers

Uploads a ParConfig and returns the ID.

uploadParConfig(par.config, user.email = NULL, as.default = FALSE)

Arguments

par.config
[ParConfig] The Parameter Configuration

user.email
[character(1)] Your email to identify yourself to the server. Does not have to be a valid one, but this identifier makes it easier to find your own submissions or submissions by colleagues. In the future we might switch to an authentication system. It could be advantageous to supply a working email if you want to migrate your submissions then.

as.default
[logical(1)] Do you want to suggest this ParConfig to be the default for the given learner in the online database. This feature is not completly implemented.

Value

[character]

Examples

par.set = makeParamSet(
  makeIntegerParam(
    id = "mtry",
    lower = expression(floor(p^0.25)),
    upper = expression(ceiling(p^0.75)),
    default = expression(round(p^0.5))),
  keys = "p")
par.config = makeParConfig(
  par.set = par.set,
  par.vals = list(ntree = 200),
  learner.name = "randomForest"
  )
if (FALSE) {
id = uploadParConfig(par.config, "jon.doe@example.com")
print(id)
}