Tune Hyperparameters for a machine learning task

Tunes the Hyperparameters for a given task and learner. Tries to find the best parameter set to tune for the given learner.

hyperopt(task, learner = NULL, par.config = NULL, hyper.control = NULL,
  show.info = getMlrOptions()$show.info)

Arguments

task

[Task] The Task

learner

[Learner] The learner that is subject to the Hyperparameter Tuning. If no learner is given the learner referenced in the par.config will be used, if available.

par.config

[ParConfig] The Parameter Configuration

hyper.control

[HyperControl] The Hyperparameter Control Object

show.info

[logical(1)] Print verbose output on console? Default is set via configureMlr.

Value

[TuneResult]

Examples

# the shortest way of hyperparameter optimization
hyperopt(iris.task, "classif.svm")

#> [Tune] Started tuning learner classif.svm for parameter set:

#>          Type len Def    Constr Req Tunable Trafo
#> cost  numeric   -   0 -15 to 15   -    TRUE     Y
#> gamma numeric   -  -2 -15 to 15   -    TRUE     Y

#> With control class: TuneControlMBO

#> Imputation value: 1Imputation value: InfImputation value: Inf

#> [Tune-x] 1: cost=48.3; gamma=175

#> [Tune-y] 1: mmce.test.mean=0.673,timetrain.test.mean=0.0053,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 2: cost=477; gamma=1.01e+03

#> [Tune-y] 2: mmce.test.mean=0.767,timetrain.test.mean=0.0047,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 3: cost=0.00167; gamma=0.00636

#> [Tune-y] 3: mmce.test.mean=0.767,timetrain.test.mean=0.0046,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 4: cost=0.178; gamma=0.00069

#> [Tune-y] 4: mmce.test.mean=0.767,timetrain.test.mean=0.0046,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 5: cost=2.38; gamma=6.49

#> [Tune-y] 5: mmce.test.mean=0.0733,timetrain.test.mean=0.005,timepredict.test.mean=0.0013; time: 0.0 min

#> [Tune-x] 6: cost=3.21e+03; gamma=0.268

#> [Tune-y] 6: mmce.test.mean=0.08,timetrain.test.mean=0.0036,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 7: cost=0.000184; gamma=5.31e-05

#> [Tune-y] 7: mmce.test.mean=0.767,timetrain.test.mean=0.0044,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 8: cost=0.0131; gamma=5.49e+03

#> [Tune-y] 8: mmce.test.mean=0.78,timetrain.test.mean=0.0041,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 9: cost=1.25e+03; gamma=1.53

#> [Tune-y] 9: mmce.test.mean=0.06,timetrain.test.mean=0.0045,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 10: cost=0.22; gamma=0.855

#> [Tune-y] 10: mmce.test.mean=0.0467,timetrain.test.mean=0.0038,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 11: cost=5.23e-05; gamma=3.15

#> [Tune-y] 11: mmce.test.mean=0.767,timetrain.test.mean=0.0048,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 12: cost=6.39; gamma=0.858

#> [Tune-y] 12: mmce.test.mean=0.0667,timetrain.test.mean=0.0044,timepredict.test.mean=0.0007; time: 0.0 min

#> [Tune-x] 13: cost=0.762; gamma=2

#> [Tune-y] 13: mmce.test.mean=0.0533,timetrain.test.mean=0.004,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 14: cost=0.976; gamma=1.06

#> [Tune-y] 14: mmce.test.mean=0.0533,timetrain.test.mean=0.0048,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 15: cost=3.28e+04; gamma=0.825

#> [Tune-y] 15: mmce.test.mean=0.0733,timetrain.test.mean=0.0035,timepredict.test.mean=0.0007; time: 0.0 min

#> [Tune-x] 16: cost=1.02e+04; gamma=4.53

#> [Tune-y] 16: mmce.test.mean=0.0667,timetrain.test.mean=0.0045,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 17: cost=0.403; gamma=0.408

#> [Tune-y] 17: mmce.test.mean=0.0467,timetrain.test.mean=0.0035,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 18: cost=161; gamma=3.64

#> [Tune-y] 18: mmce.test.mean=0.0667,timetrain.test.mean=0.0058,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 19: cost=0.399; gamma=0.621

#> [Tune-y] 19: mmce.test.mean=0.0467,timetrain.test.mean=0.0035,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 20: cost=0.288; gamma=4.08

#> [Tune-y] 20: mmce.test.mean=0.0867,timetrain.test.mean=0.0044,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 21: cost=5.65e+03; gamma=2.27

#> [Tune-y] 21: mmce.test.mean=0.0667,timetrain.test.mean=0.004,timepredict.test.mean=0.001; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 22: cost=3.28e+04; gamma=3.05e-05

#> [Tune-y] 22: mmce.test.mean=0.04,timetrain.test.mean=0.003,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 23: cost=3.27e+04; gamma=0.0229

#> [Tune-y] 23: mmce.test.mean=0.0533,timetrain.test.mean=0.0045,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 24: cost=3.27e+04; gamma=0.000188

#> [Tune-y] 24: mmce.test.mean=0.0333,timetrain.test.mean=0.0048,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 25: cost=3.27e+04; gamma=0.00209

#> [Tune-y] 25: mmce.test.mean=0.0267,timetrain.test.mean=0.0045,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune] Result: cost=3.27e+04; gamma=0.00209 : mmce.test.mean=0.0267

#> Tune result:
#> Op. pars: cost=3.27e+04; gamma=0.00209
#> mmce.test.mean=0.0267

# manually defining the paramer space configuration
par.config = makeParConfig(
  par.set = makeParamSet(
    makeIntegerParam("mtry", lower = 1, upper = 4),
    makeDiscreteParam("ntree", values = c(10, 25, 50))
  ),
  par.vals = list(replace = FALSE),
  learner.name = "randomForest"
)
hyperopt(bh.task, par.config = par.config)

#> [Tune] Started tuning learner regr.randomForest for parameter set:

#>           Type len Def   Constr Req Tunable Trafo
#> mtry   integer   -   -   1 to 4   -    TRUE     -
#> ntree discrete   -   - 10,25,50   -    TRUE     -

#> With control class: TuneControlGrid

#> Imputation value: InfImputation value: InfImputation value: Inf

#> [Tune-x] 1: mtry=1; ntree=10

#> [Tune-y] 1: mse.test.mean=25.4,timetrain.test.mean=0.003,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 2: mtry=2; ntree=10

#> [Tune-y] 2: mse.test.mean=18.6,timetrain.test.mean=0.0052,timepredict.test.mean=0.0006; time: 0.0 min

#> [Tune-x] 3: mtry=3; ntree=10

#> [Tune-y] 3: mse.test.mean=  17,timetrain.test.mean=0.0068,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 4: mtry=4; ntree=10

#> [Tune-y] 4: mse.test.mean=16.3,timetrain.test.mean=0.0088,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 5: mtry=1; ntree=25

#> [Tune-y] 5: mse.test.mean=26.4,timetrain.test.mean=0.0066,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 6: mtry=2; ntree=25

#> [Tune-y] 6: mse.test.mean=16.5,timetrain.test.mean=0.0124,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 7: mtry=3; ntree=25

#> [Tune-y] 7: mse.test.mean=15.3,timetrain.test.mean=0.014,timepredict.test.mean=0.0012; time: 0.0 min

#> [Tune-x] 8: mtry=4; ntree=25

#> [Tune-y] 8: mse.test.mean=13.4,timetrain.test.mean=0.018,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 9: mtry=1; ntree=50

#> [Tune-y] 9: mse.test.mean=24.4,timetrain.test.mean=0.014,timepredict.test.mean=0.0016; time: 0.0 min

#> [Tune-x] 10: mtry=2; ntree=50

#> [Tune-y] 10: mse.test.mean=16.8,timetrain.test.mean=0.0212,timepredict.test.mean=0.0012; time: 0.0 min

#> [Tune-x] 11: mtry=3; ntree=50

#> [Tune-y] 11: mse.test.mean=14.4,timetrain.test.mean=0.0272,timepredict.test.mean=0.0016; time: 0.0 min

#> [Tune-x] 12: mtry=4; ntree=50

#> [Tune-y] 12: mse.test.mean=12.5,timetrain.test.mean=0.0334,timepredict.test.mean=0.0012; time: 0.0 min

#> [Tune] Result: mtry=4; ntree=50 : mse.test.mean=12.5,timetrain.test.mean=0.0334,timepredict.test.mean=0.0012

#> Tune result:
#> Op. pars: mtry=4; ntree=50
#> mse.test.mean=12.5,timetrain.test.mean=0.0334,timepredict.test.mean=0.0012