Tune Hyperparameters for a machine learning task

Tunes the Hyperparameters for a given task and learner. Tries to find the best parameter set to tune for the given learner.

hyperopt(task, learner = NULL, par.config = NULL, hyper.control = NULL,
  show.info = getMlrOptions()$show.info)

Arguments

task

[Task] The Task

learner

[Learner] The learner that is subject to the Hyperparameter Tuning. If no learner is given the learner referenced in the par.config will be used, if available.

par.config

[ParConfig] The Parameter Configuration

hyper.control

[HyperControl] The Hyperparameter Control Object

show.info

[logical(1)] Print verbose output on console? Default is set via configureMlr.

Value

[TuneResult]

Examples

# the shortest way of hyperparameter optimization
hyperopt(iris.task, "classif.svm")

#> [Tune] Started tuning learner classif.svm for parameter set:

#>          Type len Def    Constr Req Tunable Trafo
#> cost  numeric   -   0 -15 to 15   -    TRUE     Y
#> gamma numeric   -  -2 -15 to 15   -    TRUE     Y

#> With control class: TuneControlMBO

#> Imputation value: 1Imputation value: InfImputation value: Inf

#> [Tune-x] 1: cost=1.91e+04; gamma=658

#> [Tune-y] 1: mmce.test.mean=0.78,timetrain.test.mean=0.0049,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 2: cost=0.0305; gamma=6.02

#> [Tune-y] 2: mmce.test.mean=0.78,timetrain.test.mean=0.0047,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 3: cost=9.93e-05; gamma=0.000411

#> [Tune-y] 3: mmce.test.mean=0.773,timetrain.test.mean=0.0049,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 4: cost=0.429; gamma=0.0751

#> [Tune-y] 4: mmce.test.mean=0.0467,timetrain.test.mean=0.0038,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 5: cost=95.5; gamma=5.29e+03

#> [Tune-y] 5: mmce.test.mean=0.78,timetrain.test.mean=0.0043,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 6: cost=6.55; gamma=0.0001

#> [Tune-y] 6: mmce.test.mean=0.72,timetrain.test.mean=0.0051,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 7: cost=1.76e+03; gamma=0.0402

#> [Tune-y] 7: mmce.test.mean=0.06,timetrain.test.mean=0.0033,timepredict.test.mean=0.0007; time: 0.0 min

#> [Tune-x] 8: cost=0.000686; gamma=71.4

#> [Tune-y] 8: mmce.test.mean= 0.8,timetrain.test.mean=0.0052,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 9: cost=3.27e+04; gamma=0.0882

#> [Tune-y] 9: mmce.test.mean=0.107,timetrain.test.mean=0.0042,timepredict.test.mean=0.0007; time: 0.0 min

#> [Tune-x] 10: cost=0.000772; gamma=0.0457

#> [Tune-y] 10: mmce.test.mean=0.773,timetrain.test.mean=0.0048,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 11: cost=15.4; gamma=0.0828

#> [Tune-y] 11: mmce.test.mean=0.0533,timetrain.test.mean=0.0031,timepredict.test.mean=0.0016; time: 0.0 min

#> [Tune-x] 12: cost=2.09; gamma=0.0519

#> [Tune-y] 12: mmce.test.mean=0.0467,timetrain.test.mean=0.0035,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 13: cost=1.48; gamma=0.108

#> [Tune-y] 13: mmce.test.mean=0.0533,timetrain.test.mean=0.0037,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 14: cost=424; gamma=0.0806

#> [Tune-y] 14: mmce.test.mean=0.0667,timetrain.test.mean=0.004,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 15: cost=0.983; gamma=0.0713

#> [Tune-y] 15: mmce.test.mean=0.0467,timetrain.test.mean=0.0047,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 16: cost=104; gamma=0.0192

#> [Tune-y] 16: mmce.test.mean=0.06,timetrain.test.mean=0.0032,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 17: cost=79.6; gamma=0.0424

#> [Tune-y] 17: mmce.test.mean=0.06,timetrain.test.mean=0.0042,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 18: cost=3.27e+04; gamma=0.00953

#> [Tune-y] 18: mmce.test.mean=0.0667,timetrain.test.mean=0.0036,timepredict.test.mean=0.0006; time: 0.0 min

#> [Tune-x] 19: cost=8.49e+03; gamma=0.018

#> [Tune-y] 19: mmce.test.mean=0.0667,timetrain.test.mean=0.0033,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 20: cost=0.494; gamma=0.0136

#> [Tune-y] 20: mmce.test.mean=0.127,timetrain.test.mean=0.0041,timepredict.test.mean=0.0011; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 21: cost=1.2e+03; gamma=0.00688

#> [Tune-y] 21: mmce.test.mean=0.06,timetrain.test.mean=0.003,timepredict.test.mean=0.0009; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 22: cost=774; gamma=0.0143

#> [Tune-y] 22: mmce.test.mean=0.0667,timetrain.test.mean=0.003,timepredict.test.mean=0.0009; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 23: cost=3.13; gamma=0.0734

#> [Tune-y] 23: mmce.test.mean=0.0467,timetrain.test.mean=0.0032,timepredict.test.mean=0.0007; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 24: cost=3.27e+04; gamma=0.00212

#> [Tune-y] 24: mmce.test.mean=0.0533,timetrain.test.mean=0.0049,timepredict.test.mean=0.0008; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 25: cost=1.75e+04; gamma=0.00401

#> [Tune-y] 25: mmce.test.mean=0.06,timetrain.test.mean=0.0034,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune] Result: cost=0.983; gamma=0.0713 : mmce.test.mean=0.0467

#> Tune result:
#> Op. pars: cost=0.983; gamma=0.0713
#> mmce.test.mean=0.0467

# manually defining the paramer space configuration
par.config = makeParConfig(
  par.set = makeParamSet(
    makeIntegerParam("mtry", lower = 1, upper = 4),
    makeDiscreteParam("ntree", values = c(10, 25, 50))
  ),
  par.vals = list(replace = FALSE),
  learner.name = "randomForest"
)
hyperopt(bh.task, par.config = par.config)

#> [Tune] Started tuning learner regr.randomForest for parameter set:

#>           Type len Def   Constr Req Tunable Trafo
#> mtry   integer   -   -   1 to 4   -    TRUE     -
#> ntree discrete   -   - 10,25,50   -    TRUE     -

#> With control class: TuneControlGrid

#> Imputation value: InfImputation value: InfImputation value: Inf

#> [Tune-x] 1: mtry=1; ntree=10

#> [Tune-y] 1: mse.test.mean=19.4,timetrain.test.mean=0.0032,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 2: mtry=2; ntree=10

#> [Tune-y] 2: mse.test.mean=16.5,timetrain.test.mean=0.005,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 3: mtry=3; ntree=10

#> [Tune-y] 3: mse.test.mean=12.3,timetrain.test.mean=0.007,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 4: mtry=4; ntree=10

#> [Tune-y] 4: mse.test.mean=12.1,timetrain.test.mean=0.0083,timepredict.test.mean=0.0006; time: 0.0 min

#> [Tune-x] 5: mtry=1; ntree=25

#> [Tune-y] 5: mse.test.mean=20.6,timetrain.test.mean=0.0078,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 6: mtry=2; ntree=25

#> [Tune-y] 6: mse.test.mean=13.3,timetrain.test.mean=0.0112,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 7: mtry=3; ntree=25

#> [Tune-y] 7: mse.test.mean=  11,timetrain.test.mean=0.0125,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 8: mtry=4; ntree=25

#> [Tune-y] 8: mse.test.mean=10.9,timetrain.test.mean=0.0161,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 9: mtry=1; ntree=50

#> [Tune-y] 9: mse.test.mean=19.7,timetrain.test.mean=0.0143,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 10: mtry=2; ntree=50

#> [Tune-y] 10: mse.test.mean=  13,timetrain.test.mean=0.022,timepredict.test.mean=0.0011; time: 0.0 min

#> [Tune-x] 11: mtry=3; ntree=50

#> [Tune-y] 11: mse.test.mean=11.1,timetrain.test.mean=0.0263,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 12: mtry=4; ntree=50

#> [Tune-y] 12: mse.test.mean=10.7,timetrain.test.mean=0.0304,timepredict.test.mean=0.0022; time: 0.0 min

#> [Tune] Result: mtry=4; ntree=50 : mse.test.mean=10.7,timetrain.test.mean=0.0304,timepredict.test.mean=0.0022

#> Tune result:
#> Op. pars: mtry=4; ntree=50
#> mse.test.mean=10.7,timetrain.test.mean=0.0304,timepredict.test.mean=0.0022