Tune Hyperparameters for a machine learning task

Tunes the Hyperparameters for a given task and learner. Tries to find the best parameter set to tune for the given learner.

hyperopt(task, learner = NULL, par.config = NULL, hyper.control = NULL,
  show.info = getMlrOptions()$show.info)

Arguments

task
[Task] The Task

learner
[Learner] The learner that is subject to the Hyperparameter Tuning. If no learner is given the learner referenced in the par.config will be used, if available.

par.config
[ParConfig] The Parameter Configuration

hyper.control
[HyperControl] The Hyperparameter Control Object

show.info
[logical(1)] Print verbose output on console? Default is set via configureMlr.

Value

[TuneResult]

Examples

# the shortest way of hyperparameter optimization
hyperopt(iris.task, "classif.svm")

#> [Tune] Started tuning learner classif.svm for parameter set:

#>          Type len Def    Constr Req Tunable Trafo
#> cost  numeric   -   0 -15 to 15   -    TRUE     Y
#> gamma numeric   -  -2 -15 to 15   -    TRUE     Y

#> With control class: TuneControlMBO

#> Imputation value: 1Imputation value: InfImputation value: Inf

#> [Tune-x] 1: cost=0.000429; gamma=0.285

#> [Tune-y] 1: mmce.test.mean=0.74,timetrain.test.mean=0.0183,timepredict.test.mean=0.0066; time: 0.0 min

#> [Tune-x] 2: cost=3.79; gamma=5.64

#> [Tune-y] 2: mmce.test.mean=0.0667,timetrain.test.mean=0.005,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 3: cost=338; gamma=0.000245

#> [Tune-y] 3: mmce.test.mean=0.0467,timetrain.test.mean=0.0038,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 4: cost=0.0376; gamma=278

#> [Tune-y] 4: mmce.test.mean= 0.8,timetrain.test.mean=0.0051,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 5: cost=0.564; gamma=0.00107

#> [Tune-y] 5: mmce.test.mean=0.72,timetrain.test.mean=0.005,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 6: cost=2.98e+04; gamma=131

#> [Tune-y] 6: mmce.test.mean=0.66,timetrain.test.mean=0.0055,timepredict.test.mean=0.0011; time: 0.0 min

#> [Tune-x] 7: cost=6.11e-05; gamma=2.44e+03

#> [Tune-y] 7: mmce.test.mean=0.767,timetrain.test.mean=0.0041,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 8: cost=46.1; gamma=0.0443

#> [Tune-y] 8: mmce.test.mean=0.04,timetrain.test.mean=0.0032,timepredict.test.mean=0.001; time: 0.0 min

#> Loading required package: rgenoud

#> ##  rgenoud (Version 5.7-12.4, Build Date: 2015-07-19)
#> ##  See http://sekhon.berkeley.edu/rgenoud for additional documentation.
#> ##  Please cite software as:
#> ##   Walter Mebane, Jr. and Jasjeet S. Sekhon. 2011.
#> ##   ``Genetic Optimization Using Derivatives: The rgenoud package for R.''
#> ##   Journal of Statistical Software, 42(11): 1-26. 
#> ##

#> [Tune-x] 9: cost=14.2; gamma=48.3

#> [Tune-y] 9: mmce.test.mean=0.367,timetrain.test.mean=0.0057,timepredict.test.mean=0.0015; time: 0.0 min

#> [Tune-x] 10: cost=107; gamma=0.00263

#> [Tune-y] 10: mmce.test.mean=0.0333,timetrain.test.mean=0.0046,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 11: cost=117; gamma=0.000988

#> [Tune-y] 11: mmce.test.mean=0.04,timetrain.test.mean=0.0036,timepredict.test.mean=0.0009; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 12: cost=221; gamma=8.48

#> [Tune-y] 12: mmce.test.mean=0.0867,timetrain.test.mean=0.0051,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 13: cost=62.7; gamma=3.11e-05

#> [Tune-y] 13: mmce.test.mean=0.36,timetrain.test.mean=0.005,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 14: cost=153; gamma=0.566

#> [Tune-y] 14: mmce.test.mean=0.0733,timetrain.test.mean=0.0031,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 15: cost=242; gamma=0.00948

#> [Tune-y] 15: mmce.test.mean=0.04,timetrain.test.mean=0.0032,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 16: cost=5.34; gamma=0.176

#> [Tune-y] 16: mmce.test.mean=0.04,timetrain.test.mean=0.0045,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 17: cost=596; gamma=0.00115

#> [Tune-y] 17: mmce.test.mean=0.04,timetrain.test.mean=0.003,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 18: cost=6.81; gamma=1.17

#> [Tune-y] 18: mmce.test.mean=0.0733,timetrain.test.mean=0.0034,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 19: cost=302; gamma=0.00119

#> [Tune-y] 19: mmce.test.mean=0.0467,timetrain.test.mean=0.0034,timepredict.test.mean=0.0008; time: 0.0 min

#> [Tune-x] 20: cost=116; gamma=0.0245

#> [Tune-y] 20: mmce.test.mean=0.0467,timetrain.test.mean=0.0045,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 21: cost=1.86e+03; gamma=0.000186

#> [Tune-y] 21: mmce.test.mean=0.0467,timetrain.test.mean=0.0031,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 22: cost=3.27e+04; gamma=0.000423

#> [Tune-y] 22: mmce.test.mean=0.04,timetrain.test.mean=0.0033,timepredict.test.mean=0.0009; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 23: cost=36.5; gamma=0.161

#> [Tune-y] 23: mmce.test.mean=0.0467,timetrain.test.mean=0.0037,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 24: cost=5.18e+03; gamma=0.000414

#> [Tune-y] 24: mmce.test.mean=0.04,timetrain.test.mean=0.0043,timepredict.test.mean=0.0011; time: 0.0 min

#> [Tune-x] 25: cost=899; gamma=0.000426

#> [Tune-y] 25: mmce.test.mean=0.0467,timetrain.test.mean=0.0037,timepredict.test.mean=0.001; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune] Result: cost=107; gamma=0.00263 : mmce.test.mean=0.0333

#> Tune result:
#> Op. pars: cost=107; gamma=0.00263
#> mmce.test.mean=0.0333

# manually defining the paramer space configuration
par.config = makeParConfig(
  par.set = makeParamSet(
    makeIntegerParam("mtry", lower = 1, upper = 4),
    makeDiscreteParam("ntree", values = c(10, 25, 50))
  ),
  par.vals = list(replace = FALSE),
  learner.name = "randomForest"
)
hyperopt(bh.task, par.config = par.config)

#> [Tune] Started tuning learner regr.randomForest for parameter set:

#>           Type len Def   Constr Req Tunable Trafo
#> mtry   integer   -   -   1 to 4   -    TRUE     -
#> ntree discrete   -   - 10,25,50   -    TRUE     -

#> With control class: TuneControlGrid

#> Imputation value: InfImputation value: InfImputation value: Inf

#> [Tune-x] 1: mtry=1; ntree=10

#> [Tune-y] 1: mse.test.mean=22.8,timetrain.test.mean=0.0036,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 2: mtry=2; ntree=10

#> [Tune-y] 2: mse.test.mean=17.4,timetrain.test.mean=0.0051,timepredict.test.mean=0.0006; time: 0.0 min

#> [Tune-x] 3: mtry=3; ntree=10

#> [Tune-y] 3: mse.test.mean=12.1,timetrain.test.mean=0.0059,timepredict.test.mean=0.0018; time: 0.0 min

#> [Tune-x] 4: mtry=4; ntree=10

#> [Tune-y] 4: mse.test.mean=15.5,timetrain.test.mean=0.0079,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 5: mtry=1; ntree=25

#> [Tune-y] 5: mse.test.mean=21.8,timetrain.test.mean=0.0085,timepredict.test.mean=0.0009; time: 0.0 min

#> [Tune-x] 6: mtry=2; ntree=25

#> [Tune-y] 6: mse.test.mean=14.7,timetrain.test.mean=0.0119,timepredict.test.mean=0.0011; time: 0.0 min

#> [Tune-x] 7: mtry=3; ntree=25

#> [Tune-y] 7: mse.test.mean=12.9,timetrain.test.mean=0.0158,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 8: mtry=4; ntree=25

#> [Tune-y] 8: mse.test.mean=12.5,timetrain.test.mean=0.0169,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune-x] 9: mtry=1; ntree=50

#> [Tune-y] 9: mse.test.mean=21.4,timetrain.test.mean=0.0131,timepredict.test.mean=0.0011; time: 0.0 min

#> [Tune-x] 10: mtry=2; ntree=50

#> [Tune-y] 10: mse.test.mean=13.8,timetrain.test.mean=0.022,timepredict.test.mean=0.0012; time: 0.0 min

#> [Tune-x] 11: mtry=3; ntree=50

#> [Tune-y] 11: mse.test.mean=12.3,timetrain.test.mean=0.0249,timepredict.test.mean=0.0178; time: 0.0 min

#> [Tune-x] 12: mtry=4; ntree=50

#> [Tune-y] 12: mse.test.mean=11.3,timetrain.test.mean=0.0295,timepredict.test.mean=0.001; time: 0.0 min

#> [Tune] Result: mtry=4; ntree=50 : mse.test.mean=11.3,timetrain.test.mean=0.0295,timepredict.test.mean=0.001

#> Tune result:
#> Op. pars: mtry=4; ntree=50
#> mse.test.mean=11.3,timetrain.test.mean=0.0295,timepredict.test.mean=0.001