Hyperparameter Tuning Control Object

Defines how the hyperparameter tuning should be conducted

makeHyperControl(mlr.control = NULL, resampling = NULL, measures = NULL,
  par.config = NULL)

Arguments

mlr.control
[TuneControl] Control object for search method. Also selects the optimization algorithm for tuning.

resampling
[ResampleDesc] The resampling determines how the performance is obtained during tuning.

measures
[Measure | list of Measure] Performance measure(s) to evaluate. Default is the default measure for the task, see here getDefaultMeasure.

par.config
[ParConfig] The Parameter Configuration

Value

[HyperControl]

See also

Other HyperControl: getHyperControlMeasures, getHyperControlMlrControl, getHyperControlResampling, setHyperControlMeasures, setHyperControlMlrControl, setHyperControlResampling

Examples

hyper.control = makeHyperControl(
  mlr.control = makeTuneControlRandom(maxit = 5),
  resampling = cv2,
  measures = acc
)
hyperopt(task = iris.task, learner = "classif.svm", hyper.control = hyper.control)

#> [Tune] Started tuning learner classif.svm for parameter set:

#>          Type len Def    Constr Req Tunable Trafo
#> cost  numeric   -   0 -15 to 15   -    TRUE     Y
#> gamma numeric   -  -2 -15 to 15   -    TRUE     Y

#> With control class: TuneControlRandom

#> Imputation value: -0

#> [Tune-x] 1: cost=1.74e+03; gamma=0.000271

#> Resampling: cross-validation

#> Measures:             acc       

#> [Tune-y] 1: acc.test.mean=0.9666667; time: 0.0 min

#> [Tune-x] 2: cost=85.7; gamma=0.000866

#> Resampling: cross-validation

#> Measures:             acc       

#> [Tune-y] 2: acc.test.mean=0.9666667; time: 0.0 min

#> [Tune-x] 3: cost=8.7e-05; gamma=0.0179

#> Resampling: cross-validation

#> Measures:             acc       

#> [Tune-y] 3: acc.test.mean=0.2866667; time: 0.0 min

#> [Tune-x] 4: cost=0.00639; gamma=0.00551

#> Resampling: cross-validation

#> Measures:             acc       

#> [Tune-y] 4: acc.test.mean=0.2866667; time: 0.0 min

#> [Tune-x] 5: cost=0.157; gamma=0.000162

#> Resampling: cross-validation

#> Measures:             acc       

#> [Tune-y] 5: acc.test.mean=0.2866667; time: 0.0 min

#> [Tune] Result: cost=85.7; gamma=0.000866 : acc.test.mean=0.9666667

#> Tune result:
#> Op. pars: cost=85.7; gamma=0.000866
#> acc.test.mean=0.9666667