Fuse learner with mlrHyperopt tuning.

Fuses an mlr base learner with mlrHyperopt tuning. Creates a learner object, which can be used like any other learner object. If the train function is called on it, hyperopt is invoked to select an optimal set of hyperparameter values. Finally, a model is fitted on the complete training data with these optimal hyperparameters and returned.

makeHyperoptWrapper(
  learner,
  par.config = NULL,
  hyper.control = NULL,
  show.info = getMlrOptions()$show.info
)

Arguments

learner
[Learner] Learner

par.config
[ParConfig] The Parameter Configuration

hyper.control
[HyperControl] The Hyperparameter Control Object

show.info
[logical(1)]
Print verbose output on console? Default is set via configureMlr.

Value

[Learner].

Examples

# \donttest{
task = makeClassifTask(data = iris, target = "Species")
lrn = makeLearner("classif.svm")
lrn = makeHyperoptWrapper(lrn)
mod = train(lrn, task)

#> [Tune] Started tuning learner classif.svm for parameter set:

#>          Type len Def    Constr Req Tunable Trafo
#> cost  numeric   -   0 -15 to 15   -    TRUE     Y
#> gamma numeric   -  -2 -15 to 15   -    TRUE     Y

#> With control class: TuneControlMBO

#> Imputation value: 1

#> [Tune-x] 1: cost=0.186; gamma=0.0461

#> [Tune-y] 1: mmce.test.mean=0.1066667; time: 0.0 min

#> [Tune-x] 2: cost=48.1; gamma=0.237

#> [Tune-y] 2: mmce.test.mean=0.0533333; time: 0.0 min

#> [Tune-x] 3: cost=8.06; gamma=1.87e+03

#> [Tune-y] 3: mmce.test.mean=0.7933333; time: 0.0 min

#> [Tune-x] 4: cost=4.53e+03; gamma=115

#> [Tune-y] 4: mmce.test.mean=0.6666667; time: 0.0 min

#> [Tune-x] 5: cost=0.053; gamma=3.07

#> [Tune-y] 5: mmce.test.mean=0.6600000; time: 0.0 min

#> [Tune-x] 6: cost=0.00264; gamma=0.000251

#> [Tune-y] 6: mmce.test.mean=0.7933333; time: 0.0 min

#> [Tune-x] 7: cost=435; gamma=0.00117

#> [Tune-y] 7: mmce.test.mean=0.0333333; time: 0.0 min

#> [Tune-x] 8: cost=6.16e-05; gamma=3.71e+03

#> [Tune-y] 8: mmce.test.mean=0.8133333; time: 0.0 min

#> [Tune-x] 9: cost=597; gamma=0.0213

#> [Tune-y] 9: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 10: cost=17.6; gamma=0.0279

#> [Tune-y] 10: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 11: cost=3.28e+04; gamma=0.00257

#> [Tune-y] 11: mmce.test.mean=0.0466667; time: 0.0 min

#> [Tune-x] 12: cost=199; gamma=0.0802

#> [Tune-y] 12: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 13: cost=303; gamma=0.00413

#> [Tune-y] 13: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 14: cost=2.7e+03; gamma=0.00163

#> [Tune-y] 14: mmce.test.mean=0.0333333; time: 0.0 min

#> [Tune-x] 15: cost=3.25e+04; gamma=0.204

#> [Tune-y] 15: mmce.test.mean=0.0666667; time: 0.0 min

#> [Tune-x] 16: cost=3.27e+04; gamma=0.000885

#> [Tune-y] 16: mmce.test.mean=0.0466667; time: 0.0 min

#> [Tune-x] 17: cost=111; gamma=0.0114

#> [Tune-y] 17: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 18: cost=3.27e+04; gamma=3.05e-05

#> [Tune-y] 18: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 19: cost=1.7e+03; gamma=8.81e-05

#> [Tune-y] 19: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 20: cost=970; gamma=3.05e-05

#> [Tune-y] 20: mmce.test.mean=0.0333333; time: 0.0 min

#> [Tune-x] 21: cost=3.94; gamma=0.105

#> [Tune-y] 21: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 22: cost=5.39e+03; gamma=3.05e-05

#> [Tune-y] 22: mmce.test.mean=0.0333333; time: 0.0 min

#> [Tune-x] 23: cost=1.22e+03; gamma=0.000409

#> [Tune-y] 23: mmce.test.mean=0.0333333; time: 0.0 min

#> [Tune-x] 24: cost=43.6; gamma=0.0527

#> [Tune-y] 24: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 25: cost=864; gamma=0.00204

#> [Tune-y] 25: mmce.test.mean=0.0400000; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune] Result: cost=5.39e+03; gamma=3.05e-05 : mmce.test.mean=0.0333333

print(getTuneResult(mod))

#> Tune result:
#> Op. pars: cost=5.39e+03; gamma=3.05e-05
#> mmce.test.mean=0.0333333

# nested resampling for evaluation
# we also extract tuned hyper pars in each iteration
r = resample(lrn, task, cv3, extract = getTuneResult)

#> Resampling: cross-validation

#> Measures:             mmce      

#> [Tune] Started tuning learner classif.svm for parameter set:

#>          Type len Def    Constr Req Tunable Trafo
#> cost  numeric   -   0 -15 to 15   -    TRUE     Y
#> gamma numeric   -  -2 -15 to 15   -    TRUE     Y

#> With control class: TuneControlMBO

#> Imputation value: 1

#> [Tune-x] 1: cost=0.204; gamma=7.46e+03

#> [Tune-y] 1: mmce.test.mean=0.7300000; time: 0.0 min

#> [Tune-x] 2: cost=2e+04; gamma=0.0131

#> [Tune-y] 2: mmce.test.mean=0.0600000; time: 0.0 min

#> [Tune-x] 3: cost=5.02; gamma=1.28

#> [Tune-y] 3: mmce.test.mean=0.0600000; time: 0.0 min

#> [Tune-x] 4: cost=5.44e-05; gamma=55.8

#> [Tune-y] 4: mmce.test.mean=0.7200000; time: 0.0 min

#> [Tune-x] 5: cost=1.33e+03; gamma=0.606

#> [Tune-y] 5: mmce.test.mean=0.0700000; time: 0.0 min

#> [Tune-x] 6: cost=0.00408; gamma=0.00174

#> [Tune-y] 6: mmce.test.mean=0.7100000; time: 0.0 min

#> [Tune-x] 7: cost=121; gamma=3.14e-05

#> [Tune-y] 7: mmce.test.mean=0.3700000; time: 0.0 min

#> [Tune-x] 8: cost=0.0124; gamma=1.48e+03

#> [Tune-y] 8: mmce.test.mean=0.7200000; time: 0.0 min

#> [Tune-x] 9: cost=59.5; gamma=0.119

#> [Tune-y] 9: mmce.test.mean=0.0600000; time: 0.0 min

#> [Tune-x] 10: cost=3.27e+04; gamma=0.116

#> [Tune-y] 10: mmce.test.mean=0.1200000; time: 0.0 min

#> [Tune-x] 11: cost=48.9; gamma=1.12

#> [Tune-y] 11: mmce.test.mean=0.0800000; time: 0.0 min

#> [Tune-x] 12: cost=1.71e+03; gamma=0.0244

#> [Tune-y] 12: mmce.test.mean=0.0700000; time: 0.0 min

#> [Tune-x] 13: cost=17.5; gamma=0.451

#> [Tune-y] 13: mmce.test.mean=0.0900000; time: 0.0 min

#> [Tune-x] 14: cost=754; gamma=0.0601

#> [Tune-y] 14: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 15: cost=646; gamma=0.0628

#> [Tune-y] 15: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 16: cost=560; gamma=0.0709

#> [Tune-y] 16: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 17: cost=3.27e+04; gamma=0.00117

#> [Tune-y] 17: mmce.test.mean=0.0700000; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 18: cost=840; gamma=8.91

#> [Tune-y] 18: mmce.test.mean=0.1200000; time: 0.0 min

#> [Tune-x] 19: cost=2.26; gamma=0.107

#> [Tune-y] 19: mmce.test.mean=0.0600000; time: 0.0 min

#> [Tune-x] 20: cost=697; gamma=0.134

#> [Tune-y] 20: mmce.test.mean=0.0800000; time: 0.0 min

#> [Tune-x] 21: cost=10; gamma=0.0698

#> [Tune-y] 21: mmce.test.mean=0.0500000; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 22: cost=9.64e+03; gamma=2.48

#> [Tune-y] 22: mmce.test.mean=0.0700000; time: 0.0 min

#> [Tune-x] 23: cost=3.28e+04; gamma=0.00404

#> [Tune-y] 23: mmce.test.mean=0.0600000; time: 0.0 min

#> [Tune-x] 24: cost=129; gamma=0.0661

#> [Tune-y] 24: mmce.test.mean=0.0500000; time: 0.0 min

#> [Tune-x] 25: cost=421; gamma=0.057

#> [Tune-y] 25: mmce.test.mean=0.0500000; time: 0.0 min

#> [Tune] Result: cost=754; gamma=0.0601 : mmce.test.mean=0.0400000

#> [Resample] iter 1:    0.0200000 

#> [Tune] Started tuning learner classif.svm for parameter set:

#>          Type len Def    Constr Req Tunable Trafo
#> cost  numeric   -   0 -15 to 15   -    TRUE     Y
#> gamma numeric   -  -2 -15 to 15   -    TRUE     Y

#> With control class: TuneControlMBO

#> Imputation value: 1

#> [Tune-x] 1: cost=0.0607; gamma=67.3

#> [Tune-y] 1: mmce.test.mean=0.7000000; time: 0.0 min

#> [Tune-x] 2: cost=134; gamma=0.0652

#> [Tune-y] 2: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 3: cost=1.08e+04; gamma=0.347

#> [Tune-y] 3: mmce.test.mean=0.1000000; time: 0.0 min

#> [Tune-x] 4: cost=0.589; gamma=1.81e+04

#> [Tune-y] 4: mmce.test.mean=0.6800000; time: 0.0 min

#> [Tune-x] 5: cost=0.000538; gamma=0.00437

#> [Tune-y] 5: mmce.test.mean=0.7000000; time: 0.0 min

#> [Tune-x] 6: cost=4.49e-05; gamma=1.29

#> [Tune-y] 6: mmce.test.mean=0.7000000; time: 0.0 min

#> [Tune-x] 7: cost=1.72; gamma=0.000215

#> [Tune-y] 7: mmce.test.mean=0.7000000; time: 0.0 min

#> [Tune-x] 8: cost=491; gamma=661

#> [Tune-y] 8: mmce.test.mean=0.6800000; time: 0.0 min

#> [Tune-x] 9: cost=3.51e+03; gamma=0.0645

#> [Tune-y] 9: mmce.test.mean=0.0500000; time: 0.0 min

#> [Tune-x] 10: cost=509; gamma=0.126

#> [Tune-y] 10: mmce.test.mean=0.0600000; time: 0.0 min

#> [Tune-x] 11: cost=319; gamma=0.0648

#> [Tune-y] 11: mmce.test.mean=0.0400000; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 12: cost=204; gamma=0.0324

#> [Tune-y] 12: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 13: cost=3.27e+04; gamma=0.0387

#> [Tune-y] 13: mmce.test.mean=0.0500000; time: 0.0 min

#> [Tune-x] 14: cost=204; gamma=0.0457

#> [Tune-y] 14: mmce.test.mean=0.0400000; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 15: cost=1.92e+04; gamma=0.0134

#> [Tune-y] 15: mmce.test.mean=0.0500000; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 16: cost=1.09e+04; gamma=0.0235

#> [Tune-y] 16: mmce.test.mean=0.0500000; time: 0.0 min

#> [Tune-x] 17: cost=225; gamma=0.0177

#> [Tune-y] 17: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 18: cost=201; gamma=0.0597

#> [Tune-y] 18: mmce.test.mean=0.0400000; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 19: cost=653; gamma=0.0195

#> [Tune-y] 19: mmce.test.mean=0.0300000; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 20: cost=1.02e+03; gamma=0.0195

#> [Tune-y] 20: mmce.test.mean=0.0300000; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 21: cost=829; gamma=0.0198

#> [Tune-y] 21: mmce.test.mean=0.0300000; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 22: cost=18.4; gamma=0.0584

#> [Tune-y] 22: mmce.test.mean=0.0300000; time: 0.0 min

#> [Tune-x] 23: cost=31.1; gamma=0.0564

#> [Tune-y] 23: mmce.test.mean=0.0400000; time: 0.0 min

#> Warning: Stopped because hard maximum generation limit was hit.

#> [Tune-x] 24: cost=6.06; gamma=0.0682

#> [Tune-y] 24: mmce.test.mean=0.0300000; time: 0.0 min

#> [Tune-x] 25: cost=11.3; gamma=0.0645

#> [Tune-y] 25: mmce.test.mean=0.0300000; time: 0.0 min

#> [Tune] Result: cost=829; gamma=0.0198 : mmce.test.mean=0.0300000

#> [Resample] iter 2:    0.0600000 

#> [Tune] Started tuning learner classif.svm for parameter set:

#>          Type len Def    Constr Req Tunable Trafo
#> cost  numeric   -   0 -15 to 15   -    TRUE     Y
#> gamma numeric   -  -2 -15 to 15   -    TRUE     Y

#> With control class: TuneControlMBO

#> Imputation value: 1

#> [Tune-x] 1: cost=150; gamma=173

#> [Tune-y] 1: mmce.test.mean=0.6900000; time: 0.0 min

#> [Tune-x] 2: cost=0.00523; gamma=0.000217

#> [Tune-y] 2: mmce.test.mean=0.7100000; time: 0.0 min

#> [Tune-x] 3: cost=0.0386; gamma=0.275

#> [Tune-y] 3: mmce.test.mean=0.3800000; time: 0.0 min

#> [Tune-x] 4: cost=10.7; gamma=191

#> [Tune-y] 4: mmce.test.mean=0.7000000; time: 0.0 min

#> [Tune-x] 5: cost=0.842; gamma=4.76

#> [Tune-y] 5: mmce.test.mean=0.0600000; time: 0.0 min

#> [Tune-x] 6: cost=359; gamma=0.00235

#> [Tune-y] 6: mmce.test.mean=0.0300000; time: 0.0 min

#> [Tune-x] 7: cost=4.73e+03; gamma=8.23e+03

#> [Tune-y] 7: mmce.test.mean=0.7300000; time: 0.0 min

#> [Tune-x] 8: cost=0.000241; gamma=0.0134

#> [Tune-y] 8: mmce.test.mean=0.7100000; time: 0.0 min

#> [Tune-x] 9: cost=3.25e+04; gamma=0.00172

#> [Tune-y] 9: mmce.test.mean=0.0700000; time: 0.0 min

#> [Tune-x] 10: cost=0.346; gamma=2.64

#> [Tune-y] 10: mmce.test.mean=0.0800000; time: 0.0 min

#> [Tune-x] 11: cost=3.27e+04; gamma=3.64

#> [Tune-y] 11: mmce.test.mean=0.0700000; time: 0.0 min

#> [Tune-x] 12: cost=3.27e+04; gamma=0.00272

#> [Tune-y] 12: mmce.test.mean=0.0600000; time: 0.0 min

#> [Tune-x] 13: cost=219; gamma=1.45

#> [Tune-y] 13: mmce.test.mean=0.0700000; time: 0.0 min

#> [Tune-x] 14: cost=1.88e+03; gamma=9.03

#> [Tune-y] 14: mmce.test.mean=0.1100000; time: 0.0 min

#> [Tune-x] 15: cost=1.57e+03; gamma=0.0114

#> [Tune-y] 15: mmce.test.mean=0.0500000; time: 0.0 min

#> [Tune-x] 16: cost=1.57e+03; gamma=0.00366

#> [Tune-y] 16: mmce.test.mean=0.0400000; time: 0.0 min

#> [Tune-x] 17: cost=25.1; gamma=3.33

#> [Tune-y] 17: mmce.test.mean=0.0600000; time: 0.0 min

#> [Tune-x] 18: cost=3.27e+04; gamma=0.0629

#> [Tune-y] 18: mmce.test.mean=0.0800000; time: 0.0 min

#> [Tune-x] 19: cost=2.89e+03; gamma=3.05e-05

#> [Tune-y] 19: mmce.test.mean=0.0500000; time: 0.0 min

#> [Tune-x] 20: cost=522; gamma=0.000133

#> [Tune-y] 20: mmce.test.mean=0.0600000; time: 0.0 min

#> [Tune-x] 21: cost=3.05e-05; gamma=8.34

#> [Tune-y] 21: mmce.test.mean=0.6800000; time: 0.0 min

#> [Tune-x] 22: cost=152; gamma=3.05e-05

#> [Tune-y] 22: mmce.test.mean=0.2400000; time: 0.0 min

#> [Tune-x] 23: cost=9.02e+03; gamma=0.000123

#> [Tune-y] 23: mmce.test.mean=0.0300000; time: 0.0 min

#> [Tune-x] 24: cost=2.46e+03; gamma=0.000347

#> [Tune-y] 24: mmce.test.mean=0.0300000; time: 0.0 min

#> [Tune-x] 25: cost=3.28e+04; gamma=3.05e-05

#> [Tune-y] 25: mmce.test.mean=0.0300000; time: 0.0 min

#> [Tune] Result: cost=9.02e+03; gamma=0.000123 : mmce.test.mean=0.0300000

#> [Resample] iter 3:    0.0400000 

#> 

#> Aggregated Result: mmce.test.mean=0.0400000

#> 

getNestedTuneResultsX(r)

#>        cost        gamma
#> 1  754.1959 0.0600704428
#> 2  829.0738 0.0198470809
#> 3 9020.1068 0.0001232661

# }